refactor: moved fuunction to context
